+++ palmod/Game/Game_DevMode_A.cpp
@@ -467,7 +467,7 @@ void CGame_DevMode_A::LoadSpecificPaletteData(uint32_t nUnitId, uint32_t nPalId)
             cbPaletteSizeOnDisc = static_cast<int>(max(0, (paletteData->nPaletteOffsetEnd - paletteData->nPaletteOffset)));
 
             m_nCurrentPaletteROMLocation = paletteData->nPaletteOffset;
-            m_nCurrentPaletteSizeInColors = cbPaletteSizeOnDisc / m_nSizeOfColorsInBytes;
+            m_nCurrentPaletteSizeInColors = static_cast<uint16_t>(cbPaletteSizeOnDisc / m_nSizeOfColorsInBytes);
             m_pszCurrentPaletteName = paletteData->szPaletteName;
         }
         else
