+++ palmod/Game/Game_SFA3_A.cpp
@@ -281,7 +281,7 @@ BOOL CGame_SFA3_Core::UpdatePalImg(int Node01, int Node02, int Node03, int Node0
                                 break;
                             default:
                                 // Anything past this just gets default pairing
-                                vnPeerPaletteDistances.push_back(nPairIndex);
+                                vnPeerPaletteDistances.push_back(static_cast<int8_t>(nPairIndex));
                                 break;
                         }
